+++ packages/stream_chat/CHANGELOG.md
@@ -1,3 +1,9 @@
+## Upcoming
+
+🐞 Fixed
+
+- Remove disposed channel clients from the client state.
+
 ## 5.0.0
 
 - Included the changes from version [4.5.0](#450).